+++ src/element.rs
@@ -13,6 +13,7 @@ use crate::types::{attr_split, attr_split_cycle, fstr, strp, AttrMap, ClassList,
 use core::fmt::Display;
 use lazy_regex::{regex, Captures};
 use std::collections::HashMap;
+use std::f32::consts::{FRAC_1_SQRT_2, SQRT_2};
 use std::str::FromStr;
 
 #[derive(Clone, Debug, PartialEq)]
@@ -561,7 +562,16 @@ impl SvgElement {
                 SvgdxError::ReferenceError(format!("Element {elref} not found at this time"))
             })?;
             {
-                if let Ok(Some(el_bb)) = ctx.get_element_bbox(el) {
+                let bb = if is_surround {
+                    ctx.get_element_bbox(el)
+                } else {
+                    // TODO: this doesn't handle various cases when at least one
+                    // circle/ellipses are is present and ref_list.len() > 1.
+                    // Should probably fold the list and provide next element type
+                    // as the target shape here
+                    el.inscribed_bbox(&self.name)
+                };
+                if let Ok(Some(el_bb)) = bb {
                     bbox_list.push(el_bb);
                 } else {
                     return Err(SvgdxError::InvalidData(format!(
@@ -588,11 +598,47 @@ impl SvgElement {
             }
         }
         if let Some(bb) = bbox {
-            self.position_from_bbox(&bb);
+            self.position_from_bbox(&bb, !is_surround);
         }
         self.add_class(&format!("d-{contain_str}"));
         self.remove_attrs(&["surround", "inside", "margin"]);
         Ok(())
+    }
+
+    /// Calculate bounding box of target_shape inside self
+    pub fn inscribed_bbox(&self, target_shape: &str) -> Result<Option<BoundingBox>> {
+        let zstr = "0".to_owned();
+        match (target_shape, self.name.as_str()) {
+            // rect inside circle
+            ("rect", "circle") => {
+                if let Some(r) = self.attrs.get("r") {
+                    let cx = self.attrs.get("cx").unwrap_or(&zstr);
+                    let cy = self.attrs.get("cy").unwrap_or(&zstr);
+                    let cx = strp(cx)?;
+                    let cy = strp(cy)?;
+                    let r = strp(r)? * FRAC_1_SQRT_2;
+                    Ok(Some(BoundingBox::new(cx - r, cy - r, cx + r, cy + r)))
+                } else {
+                    Ok(None)
+                }
+            }
+            // rect inside ellipse
+            ("rect", "ellipse") => {
+                if let (Some(rx), Some(ry)) = (self.attrs.get("rx"), self.attrs.get("ry")) {
+                    let cx = self.attrs.get("cx").unwrap_or(&zstr);
+                    let cy = self.attrs.get("cy").unwrap_or(&zstr);
+                    let cx = strp(cx)?;
+                    let cy = strp(cy)?;
+                    let rx = strp(rx)? * FRAC_1_SQRT_2;
+                    let ry = strp(ry)? * FRAC_1_SQRT_2;
+                    Ok(Some(BoundingBox::new(cx - rx, cy - ry, cx + rx, cy + ry)))
+                } else {
+                    Ok(None)
+                }
+            }
+            // Trivial cases: same shape
+            _ => self.bbox(),
+        }
     }
 
     pub fn bbox(&self) -> Result<Option<BoundingBox>> {
@@ -761,7 +807,7 @@ impl SvgElement {
         Ok(new_elem)
     }
 
-    fn position_from_bbox(&mut self, bb: &BoundingBox) {
+    fn position_from_bbox(&mut self, bb: &BoundingBox, inscribe: bool) {
         let width = bb.width();
         let height = bb.height();
         let (cx, cy) = bb.center();
@@ -776,14 +822,28 @@ impl SvgElement {
             "circle" => {
                 self.attrs.insert("cx", fstr(cx));
                 self.attrs.insert("cy", fstr(cy));
-                self.attrs
-                    .insert("r", fstr(0.5 * width.max(height) * 1.414));
+                let r = if inscribe {
+                    0.5 * width.min(height)
+                } else {
+                    0.5 * width.max(height) * SQRT_2
+                };
+                self.attrs.insert("r", fstr(r));
             }
             "ellipse" => {
                 self.attrs.insert("cx", fstr(cx));
                 self.attrs.insert("cy", fstr(cy));
-                self.attrs.insert("rx", fstr(0.5 * width * 1.414));
-                self.attrs.insert("ry", fstr(0.5 * height * 1.414));
+                let rx = if inscribe {
+                    0.5 * width
+                } else {
+                    0.5 * width * SQRT_2
+                };
+                let ry = if inscribe {
+                    0.5 * height
+                } else {
+                    0.5 * height * SQRT_2
+                };
+                self.attrs.insert("rx", fstr(rx));
+                self.attrs.insert("ry", fstr(ry));
             }
             _ => {}
         }